fix small, but very important bug
git-svn-id: file:///home/mroda/Software/Genie/NewGlobal/generator/trunk@7764 cc9776de-3512-45ca-aafc-e2d9ed43c22c

+++ src/Physics/Multinucleon/EventGen/MECGenerator.cxx
@@ -621,7 +621,7 @@ void MECGenerator::SelectNSVLeptonKinematics (GHepRecord * event) const
 
   // The accept/reject loop tests a rand against a maxxsec - must scale with A.
   int NuclearA = 12;
-  int NuclearAfactorXSecMax = 1.0;
+  double NuclearAfactorXSecMax = 1.0;
   if (TgtPDG != kPdgTgtC12) {
     if (TgtPDG > kPdgTgtFreeN && TgtPDG) {
       NuclearA = pdg::IonPdgCodeToA(TgtPDG);
